fix idx function

+++ examples/arkode/F2003_serial/ark_bruss1D_FEM_klu_f2003.f90
@@ -99,8 +99,8 @@ contains
 
   ! function that maps 2D data into 1D address space
   ! (0-based since CSR matrix will be sent to C solver)
-  integer(c_int) function idx(ix, ivar)
-    integer(c_int) :: ivar, ix
+  integer(kind=myindextype) function idx(ix, ivar)
+    integer(kind=myindextype) :: ivar, ix
     idx = neqreal*(ix - 1) + ivar - 1
   end function idx
